use the right quota in the GKE requests
