+++ testutils/clustermanager/e2e-tests/gke.go
@@ -24,113 +24,11 @@ import (
 
 	"github.com/davecgh/go-spew/spew"
 	container "google.golang.org/api/container/v1beta1"
+	"google.golang.org/api/option"
 
 	"knative.dev/pkg/test/gke"
-	"knative.dev/pkg/testutils/clustermanager/e2e-tests/boskos"
 	"knative.dev/pkg/testutils/clustermanager/e2e-tests/common"
 )
-
-// GKEClient implements Client
-type GKEClient struct {
-}
-
-// GKERequest contains all requests collected for cluster creation
-type GKERequest struct {
-	// Request holds settings for GKE native operations
-	gke.Request
-
-	// BackupRegions: fall back regions to try out in case of cluster creation
-	// failure due to regional issue(s)
-	BackupRegions []string
-
-	// SkipCreation: skips cluster creation
-	SkipCreation bool
-
-	// ResourceType: the boskos resource type to acquire to hold the cluster in create
-	ResourceType string
-}
-
-// GKECluster implements ClusterOperations
-type GKECluster struct {
-	Request *GKERequest
-	// Project might be GKE specific, so put it here
-	Project string
-	Cluster  *container.Cluster
-
-	// isBoskos is true if the GCP project used is managed by boskos
-	isBoskos bool
-	// asyncCleanup tells whether the cluster needs to be deleted asynchronously afterwards
-	// It should be true on Prow but false on local.
-	asyncCleanup bool
-	operations   gke.SDKOperations
-	boskosOps    boskos.Operation
-}
-
-// Setup sets up a GKECluster client, takes GEKRequest as parameter and applies
-// all defaults if not defined.
-func (gs *GKEClient) Setup(r GKERequest) ClusterOperations {
-	gc := &GKECluster{}
-
-	if r.Project != "" { // use provided project to create cluster
-		gc.Project = r.Project
-	} else if common.IsProw() { // if no project is provided and is on Prow, use boskos
-		gc.isBoskos = true
-		gc.asyncCleanup = true
-	}
-
-	if r.MinNodes == 0 {
-		r.MinNodes = defaultGKEMinNodes
-	}
-	if r.MaxNodes == 0 {
-		r.MaxNodes = defaultGKEMaxNodes
-		// We don't want MaxNodes < MinNodes
-		if r.MinNodes > r.MaxNodes {
-			r.MaxNodes = r.MinNodes
-		}
-	}
-	if r.NodeType == "" {
-		r.NodeType = defaultGKENodeType
-	}
-	// Only use default backup regions if region is not provided
-	if len(r.BackupRegions) == 0 && r.Region == "" {
-		r.BackupRegions = defaultGKEBackupRegions
-		if common.GetOSEnv(backupRegionEnv) != "" {
-			r.BackupRegions = strings.Split(common.GetOSEnv(backupRegionEnv), " ")
-		}
-	}
-	if r.Region == "" {
-		r.Region = defaultGKERegion
-		if common.GetOSEnv(regionEnv) != "" {
-			r.Region = common.GetOSEnv(regionEnv)
-		}
-	}
-	if r.Zone == "" {
-		r.Zone = defaultGKEZone
-	} else { // No backupregions if zone is provided
-		r.BackupRegions = make([]string, 0)
-	}
-
-	if r.ResourceType == "" {
-		r.ResourceType = defaultResourceType
-	}
-
-	gc.Request = &r
-
-	client, err := gke.NewSDKClient()
-	if err != nil {
-		log.Fatalf("Failed to create GKE SDK client: '%v'", err)
-	}
-	gc.operations = client
-
-	gc.boskosOps, err = boskos.NewClient("", /* boskos owner */
-		"", /* boskos user */
-		"" /* boskos password file */)
-	if err != nil {
-		log.Fatalf("Failed to create boskos client: '%v", err)
-	}
-
-	return gc
-}
 
 // Provider returns gke
 func (gc *GKECluster) Provider() string {
@@ -143,7 +41,7 @@ func (gc *GKECluster) Provider() string {
 // Region or Zone is provided then there is no retries
 func (gc *GKECluster) Acquire() error {
 	if err := gc.checkEnvironment(); err != nil {
-		return fmt.Errorf("failed checking project/cluster from environment: '%v'", err)
+		return fmt.Errorf("failed checking project/cluster from environment: '%w'", err)
 	}
 	// If gc.Cluster is discovered above, then the cluster exists and it's
 	// project and name matches with requested, use it
@@ -172,6 +70,10 @@ func (gc *GKECluster) Acquire() error {
 	gc.ensureProtected()
 	log.Printf("Identified project %s for cluster creation", gc.Project)
 
+	client, err := newGKEClient(gc.Project, gc.Request.Environment)
+	if err != nil {
+		return fmt.Errorf("failed creating the GKE client: '%w'", err)
+	}
 	// Make a deep copy of the request struct, since the original request is supposed to be immutable
 	request := gc.Request.DeepCopy()
 	// We are going to use request for creating cluster, set its Project
@@ -196,7 +98,7 @@ func (gc *GKECluster) Acquire() error {
 	var cluster *container.Cluster
 	rb, err := gke.NewCreateClusterRequest(request)
 	if err != nil {
-		return fmt.Errorf("failed building the CreateClusterRequest: '%v'", err)
+		return fmt.Errorf("failed building the CreateClusterRequest: '%w'", err)
 	}
 	for i, region := range regions {
 		// Restore innocence
@@ -204,22 +106,22 @@ func (gc *GKECluster) Acquire() error {
 
 		clusterName := request.ClusterName
 		// Use cluster if it already exists and running
-		existingCluster, _ := gc.operations.GetCluster(gc.Project, region, request.Zone, clusterName)
+		existingCluster, _ := client.GetCluster(gc.Project, region, request.Zone, clusterName)
 		if existingCluster != nil && existingCluster.Status == clusterRunning {
 			gc.Cluster = existingCluster
 			return nil
 		}
 		// Creating cluster
 		log.Printf("Creating cluster %q in region %q zone %q with:\n%+v", clusterName, region, request.Zone, spew.Sdump(rb))
-		err = gc.operations.CreateCluster(gc.Project, region, request.Zone, rb)
+		err = client.CreateCluster(gc.Project, region, request.Zone, rb)
 		if err == nil {
-			cluster, err = gc.operations.GetCluster(gc.Project, region, request.Zone, rb.Cluster.Name)
+			cluster, err = client.GetCluster(gc.Project, region, request.Zone, rb.Cluster.Name)
 		}
 		if err != nil {
 			errMsg := fmt.Sprintf("Error during cluster creation: '%v'. ", err)
 			if !common.IsProw() { // Delete half created cluster if it's user created
 				errMsg = fmt.Sprintf("%sDeleting cluster %q in region %q zone %q in background...\n", errMsg, clusterName, region, request.Zone)
-				gc.operations.DeleteClusterAsync(gc.Project, region, request.Zone, clusterName)
+				client.DeleteClusterAsync(gc.Project, region, request.Zone, clusterName)
 			}
 			// Retry another region if cluster creation failed.
 			if i != len(regions)-1 && needsRetryCreation(err.Error()) {
@@ -254,6 +156,28 @@ func (gc *GKECluster) Delete() error {
 		return fmt.Errorf("failed checking project/cluster from environment: '%w'", err)
 	}
 	gc.ensureProtected()
+
+	// Should only get here if running locally and cluster created by this
+	// client, so at this moment cluster should have been set
+	if gc.Cluster == nil {
+		return errors.New("cluster doesn't exist")
+	}
+
+	log.Printf("Deleting cluster %q in %q", gc.Cluster.Name, gc.Cluster.Location)
+	client, err := newGKEClient(gc.Project, gc.Request.Environment)
+	if err != nil {
+		return fmt.Errorf("failed creating the GKE client: '%w'", err)
+	}
+	region, zone := gke.RegionZoneFromLoc(gc.Cluster.Location)
+	if gc.asyncCleanup {
+		_, err = client.DeleteClusterAsync(gc.Project, region, zone, gc.Cluster.Name)
+	} else {
+		err = client.DeleteCluster(gc.Project, region, zone, gc.Cluster.Name)
+	}
+	if err != nil {
+		return fmt.Errorf("failed deleting cluster: '%w'", err)
+	}
+
 	// Release Boskos if running in Prow
 	if gc.isBoskos {
 		log.Printf("Releasing Boskos resource: '%v'", gc.Project)
@@ -262,21 +186,6 @@ func (gc *GKECluster) Delete() error {
 		}
 	}
 
-	// Should only get here if running locally and cluster created by this
-	// client, so at this moment cluster should have been set
-	if gc.Cluster == nil {
-		return fmt.Errorf("cluster doesn't exist")
-	}
-	log.Printf("Deleting cluster %q in %q", gc.Cluster.Name, gc.Cluster.Location)
-	region, zone := gke.RegionZoneFromLoc(gc.Cluster.Location)
-	if gc.asyncCleanup {
-		_, err = gc.operations.DeleteClusterAsync(gc.Project, region, zone, gc.Cluster.Name)
-	} else {
-		err = gc.operations.DeleteCluster(gc.Project, region, zone, gc.Cluster.Name)
-	}
-	if err != nil {
-		return fmt.Errorf("failed deleting cluster: '%w'", err)
-	}
 	return nil
 }
 
@@ -326,9 +235,13 @@ func (gc *GKECluster) checkEnvironment() error {
 				region, zone := gke.RegionZoneFromLoc(location)
 				// Use the cluster only if project and clustername match
 				if (gc.Request.Project == "" || gc.Request.Project == project) && (gc.Request.ClusterName == "" || gc.Request.ClusterName == clusterName) {
-					cluster, err := gc.operations.GetCluster(project, region, zone, clusterName)
+					client, err := newGKEClient(project, gc.Request.Environment)
 					if err != nil {
-						return fmt.Errorf("couldn't find cluster %s in %s in %s, does it exist? %v", clusterName, project, location, err)
+						return fmt.Errorf("failed creating the GKE client: '%w'", err)
+					}
+					cluster, err := client.GetCluster(project, region, zone, clusterName)
+					if err != nil {
+						return fmt.Errorf("couldn't find cluster %s in %s in %s, does it exist? %w", clusterName, project, location, err)
 					}
 					gc.Cluster = cluster
 					gc.Project = project
@@ -351,11 +264,22 @@ func (gc *GKECluster) checkEnvironment() error {
 	// if gcloud is pointing to a project, use it
 	output, err = common.StandardExec("gcloud", "config", "get-value", "project")
 	if err != nil {
-		return fmt.Errorf("failed getting gcloud project: '%v'", err)
+		return fmt.Errorf("failed getting gcloud project: '%w'", err)
 	}
 	if string(output) != "" {
 		project := strings.Trim(strings.TrimSpace(string(output)), "\n\r")
 		gc.Project = project
 	}
 	return nil
+}
+
+// newGKEClient returns a new GKE client. project and environment must be provided.
+func newGKEClient(project, environment string) (gke.SDKOperations, error) {
+	endpoint, err := gke.ServiceEndpoint(environment)
+	if err != nil {
+		return nil, err
+	}
+	return gke.NewSDKClient(
+		option.WithEndpoint(endpoint),
+		option.WithQuotaProject(project))
 }
